safari-534.53-branch/LayoutTests/http/tests/security/postMessage/target-origin.html

<!DOCTYPE html>
<html>
<head>
<script src="resources/recv.js"></script>
<script>
if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.waitUntilDone();
}

addEventListener("message", recv, false);

function tryPostMessage(win, origin) {
    try {
        win.postMessage("Trying origin=" + origin, origin);
    } catch(ex) {
        log("Error sending message to " + origin + ". " + ex);
    }
}

function test() {
    var iframe = document.getElementById('child');
    var win = iframe.contentWindow;

    // Should succeed:
    tryPostMessage(win, "http://localhost:8000");
    tryPostMessage(win, "http://localhost:8000/");
    tryPostMessage(win, "http://localhost:8000/foo");
    tryPostMessage(win, "http://localhost:8000/foo?bar");
    tryPostMessage(win, "http://localhost:8000/foo?bar#baz");
    tryPostMessage(win, "http://user:pass@localhost:8000/foo?bar#baz");
    tryPostMessage(win, "*");

    // Should fail:
    tryPostMessage(win, "http://localhost:9090");
    tryPostMessage(win, "http://localhost");
    tryPostMessage(win, "https://localhost");
    tryPostMessage(win, "https://localhost:8000");
    tryPostMessage(win, "http://www.example.com");

    // Should throw syntax error:
    tryPostMessage(win, null);
    tryPostMessage(win, undefined);

    win.postMessage('done', '*');
}
</script>
<body onload="test()">
<div>window.location.href = <script>document.write(window.location.href);</script></div>
<div><iframe src="http://localhost:8000/security/postMessage/resources/post-message-listener.html"
  id="child" width="800" height="300" style="border: 1px solid black;">
</iframe></div>
<div id="result">waiting...</div>
</body>
</html>
